Extract wp domain folder
This is where the wordpress component wrappers and
wordpress-specific components go.
