+++ client/actions/form.ts
@@ -3,3 +3,5 @@ import { createAction } from 'typesafe-actions';
 export const change = createAction('CHANGE', resolve => (value: string) =>
   resolve({ value }),
 );
+
+export const click = createAction('CLICK');
