+++ client/components/index.ts
@@ -3,6 +3,4 @@ export { default as Commits } from './Commits';
 export { default as EditPage } from './EditPage';
 export { default as SearchPopup } from './SearchPopup';
 export { default as SettingsPage } from './SettingsPage';
-export * from './Shortcode';
 export * from './StatusMapper';
-export { default as TextControl } from './TextControl';
