+++ client/wp/index.ts
@@ -1,0 +1,3 @@
+export { default as Button } from './Button';
+export * from './Shortcode';
+export { default as TextControl } from './TextControl';
